fix #6 again: use unsigned long instead of unsigned long long
Works well on everything but Windows. In Windows half of bytes still gets truncated
with 64-bit targets. Fuck it.

+++ examples/lcm.c
@@ -22,7 +22,7 @@ int main(int argc, char* argv[]) {
     } else if (b == 0) {
         fputs("Right value is required but not set\n", stderr);
     } else {
-        printf("%llu\n", (unsigned long long) (lcm(a, b)));
+        printf("%lu\n", (unsigned long) (lcm(a, b)));
     }
     qc_args_free(args);
 }
